Use long for length in <cls>_to_cstr functions

+++ src/array.c
@@ -159,7 +159,7 @@ void array_for_each(Array *self, void (*obj_cb)(void *))
 
 size_t array_to_cstr(Array *self, char *cstr, size_t size)
 {
-    int len = 0;
+    long len = 0;
     len += cstr_ncopy(cstr, "[", size);
     void *obj = self->data;
     for (int i = 0; i < self->len; i++) {
